refactor: wire entry point and test infra to new layout
Update main.go, db config/factory, and testutil suites to match
the reorganized domain/infra packages.
