+++ backend/internal/infra/db/config.go
@@ -1,4 +1,4 @@
-package database
+package db
 
 import "backend/pkg"
