+++ backend/internal/infra/db/factory.go
@@ -1,4 +1,4 @@
-package database
+package db
 
 import (
 	"errors"
