+++ backend/internal/testutil/sqlite_suite.go
@@ -1,8 +1,8 @@
 package testutil
 
 import (
-	"backend/internal/domain"
 	"backend/internal/infra/db"
+	"backend/internal/infra/db/dao"
 	"fmt"
 	"os"
 
@@ -24,7 +24,7 @@ func (suite *DBSQLiteSuite) SetupSuite() {
 	suite.Assert().Nil(err)
 	suite.DB = db
 
-	for _, model := range domain.NewDomains() {
+	for _, model := range dao.NewDAOs() {
 		err := suite.DB.AutoMigrate(model)
 		suite.Assert().Nil(err)
 	}
